Mouse Movement Analysis/requirements.txt
# Core ML/DL Libraries
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Data Processing
scipy>=1.11.0

# Deep Learning Utilities
tensorboard>=2.13.0
tqdm>=4.65.0

# API Framework
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0
python-multipart>=0.0.6

# Security & Encryption
cryptography>=41.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Monitoring & Logging
loguru>=0.7.0
prometheus-client>=0.17.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
httpx>=0.24.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Performance Metrics
scikit-metrics>=0.1.0

